feat(starrating): convert to typescript
